refactor: search logic to storage pkg
